:recycle: Move getDelaiRealisation in projectAppelOffre module
