+++ src/dataAccess/inMemory/appelOffre.ts
@@ -103,23 +103,4 @@ const getAppelOffre = (args: {
   return appelOffre
 }
 
-const getDelaiDeRealisation = (
-  appelOffreId: string,
-  technologie: string | undefined
-): number | null => {
-  const appelOffre = appelsOffreStatic.find((ao) => ao.id === appelOffreId)
-  if (!appelOffre) return null
-  if (appelOffre.decoupageParTechnologie) {
-    if (!isValidTechnologie(technologie)) return null
-    return appelOffre.delaiRealisationEnMoisParTechnologie[technologie]
-  }
-  return appelOffre.delaiRealisationEnMois
-}
-
-const isValidTechnologie = (
-  technologie: string | undefined
-): technologie is 'pv' | 'eolien' | 'hydraulique' => {
-  return !!technologie && ['pv', 'eolien', 'hydraulique'].includes(technologie)
-}
-
-export { appelOffreRepo, appelsOffreStatic, getAppelOffre, getDelaiDeRealisation }
+export { appelOffreRepo, appelsOffreStatic, getAppelOffre }
