+++ src/modules/projectAppelOffre/index.ts
@@ -1,1 +1,2 @@
+export * from './getDelaiDeRealisation'
 export * from './isSoumisAuxGFs'
